Move integrate_trajectories function
